se añadio un fragmento para el inicio de sesion y ya se puede crear players desde su respectivo fragmento

+++ app/src/main/java/com/luisro00005513/crossoversports/DAO/PlayerXTeamDAO.java
@@ -14,7 +14,7 @@ public interface PlayerXTeamDAO {
     List<PlayerXTeam> getAll();
 
     @Query("SELECT team_id FROM PlayerXTeam WHERE player_id=:playerId")
-    PlayerXTeam playerById(Integer playerId);
+    PlayerXTeam playerById(Long playerId);
 
     @Insert
     void insertAll(PlayerXTeam... playerXTeams);
